files/requirements.txt
# PDF Processing & Conversion
PyMuPDF>=1.23.0
pymupdf4llm>=0.1.0

# API & Web Requests
requests>=2.31.0
datalab-api>=0.1.0

# LlamaParse Integration
llama-parse>=0.4.0
nest_asyncio>=1.5.0

# Environment & Configuration
python-dotenv>=1.0.0
